feat: separate  from GPADATA, ref https://github.com/QCampusmate-Dev-Team/QCampusmate-Mate-4-Class/issues/17
