+++ application/extension/src/drc/App.vue
@@ -40,8 +40,8 @@
   // import { DRC } from '@qcampusmate-mate/plasapo-core'
   import { DRC } from './DRC' 
   // import { DegReqChecker, PlannerYearCollapse } from '@qcampusmate-mate/plasapo-ui'
-  import DegReqChecker from '@/components/DegReqChecker.vue' 
-  import PlannerYearCollapse from '@/components/PlannerYearCollapse.vue'
+  import DegReqChecker from '@/drc/components/DegReqChecker.vue' 
+  import PlannerYearCollapse from '@/drc/components/PlannerYearCollapse.vue'
 
   const drc = new DRC().initialize()
   provide('drc', drc)
