+++ application/extension/src/drc/components/AcademicPlannerTable.vue
