+++ application/extension/src/drc/components/HelloWorld.vue
